requirements.txt
Django==4.1.3
django-bootstrap5==22.1
django-phonenumber-field==7.0.0
phonenumberslite==8.13.0
Pillow==9.3.0
sqlparse==0.4.3
